dump hub_name if not in config.json
